Add Update Scripts

+++ InsertTranslations/Program.cs
@@ -62,55 +62,70 @@ namespace InsertTranslations
                 File.Delete(sqlAllFilePath);
 
             string[] fileEntries = Directory.GetFiles(Path.Combine(Path.GetFullPath(AppDomain.CurrentDomain.BaseDirectory), "SQL"), "*.sql");
-            StringBuilder sqlScript = new StringBuilder();
-            sqlScript.AppendLine("-------------------------- Custom Script ---------------------------");
+            StringBuilder sqlScriptFinal = new StringBuilder();
+            StringBuilder sqlScriptInsert = new StringBuilder();
+            StringBuilder sqlScriptUpdate = new StringBuilder();
+            StringBuilder stringLines = new StringBuilder();
+            sqlScriptFinal.AppendLine("-------------------------- Custom Script ---------------------------");
             foreach (string file in fileEntries)
-            {
+            {   
                 using (StreamReader sr = File.OpenText(file))
                 {
                     string s = "";
                     while ((s = sr.ReadLine()) != null)
                     {
-                        sqlScript.AppendLine(s);
+                        stringLines.AppendLine(s);
+                    }
+
+                    if (Path.GetFileName(file).Contains("Update"))
+                    {
+                        sqlScriptUpdate.Append(stringLines);
+                    }
+                    else
+                    {
+                        sqlScriptInsert.Append(stringLines);
                     }
                 }
+                stringLines.Clear();
             }
-            sqlScript.AppendLine("-------------------------- Add Custom Script Updates/Insert ---------------------------");
+            sqlScriptFinal.AppendLine("-------------------------- Add Custom Script Insert Scripts ---------------------------");
+            sqlScriptFinal.Append(sqlScriptInsert);
+            sqlScriptFinal.AppendLine("-------------------------- Add Update Scripts ---------------------------");
+            sqlScriptFinal.AppendLine("GO");
+            sqlScriptFinal.Append(sqlScriptUpdate);
+            sqlScriptFinal.AppendLine("PRINT('------------ Summary ------------')");
+            sqlScriptFinal.AppendLine("GO");
+            sqlScriptFinal.AppendLine("UPDATE st");
+            sqlScriptFinal.AppendLine("SET st.TranslatedText = fd.TranslatedText");
+            sqlScriptFinal.AppendLine("FROM X_StaticTranslations_FactoryDefaults fd");
+            sqlScriptFinal.AppendLine("LEFT JOIN X_StaticTranslations st on fd.Cd = st.Cd and fd.[Language] = st.LanguageID");
+            sqlScriptFinal.AppendLine("WHERE fd.TranslatedText <> st.TranslatedText and isnull(st.NoUpdate,0) = 0;");
+            sqlScriptFinal.AppendLine("GO");
+            sqlScriptFinal.AppendLine("\tINSERT INTO X_StaticTranslations (Cd, LanguageID, Category, TranslatedText)");
+            sqlScriptFinal.AppendLine("SELECT fd.Cd, fd.Language, fd.Category,fd.TranslatedText");
+            sqlScriptFinal.AppendLine("FROM X_StaticTranslations_FactoryDefaults fd");
+            sqlScriptFinal.AppendLine("LEFT JOIN X_StaticTranslations st on fd.Cd = st.Cd and fd.[Language] = st.LanguageID");
+            sqlScriptFinal.AppendLine("WHERE st.Cd is null;");
+            sqlScriptFinal.AppendLine("GO");
+            sqlScriptFinal.AppendLine("UPDATE o");
+            sqlScriptFinal.AppendLine("SET o.VALUE = fd.VALUE");
+            sqlScriptFinal.AppendLine("FROM L_Object_FactoryDefaults fd");
+            sqlScriptFinal.AppendLine("LEFT JOIN L_Object o on o.ID_TABLE =fd.ID_TABLE and o.ID_LANGUAGES = fd.ID_LANGUAGES and o.TABLE_NAME = fd.TABLE_NAME");
+            sqlScriptFinal.AppendLine("WHERE o.VALUE <> fd.VALUE and isnull(o.NO_UPDATE,0) = 0;");
+            sqlScriptFinal.AppendLine("GO");
+            sqlScriptFinal.AppendLine("\tINSERT INTO L_Object (ID_TABLE, ID_LANGUAGES, DATA_TYPE,VALUE,TABLE_NAME)");
+            sqlScriptFinal.AppendLine("SELECT fd.ID_TABLE, fd.ID_LANGUAGES, fd.DATA_TYPE, fd.VALUE, fd.TABLE_NAME");
+            sqlScriptFinal.AppendLine("FROM L_Object_FactoryDefaults fd");
+            sqlScriptFinal.AppendLine("LEFT JOIN L_Object l on l.ID_TABLE = fd.ID_TABLE and l.ID_LANGUAGES = fd.ID_LANGUAGES and l.TABLE_NAME = fd.TABLE_NAME");
+            sqlScriptFinal.AppendLine("WHERE l.ID_TABLE is null AND l.VALUE is null;");
+            sqlScriptFinal.AppendLine("GO");
+            sqlScriptFinal.AppendLine("PRINT 'Custom Script Completed'");
 
-            sqlScript.AppendLine("GO");
-            sqlScript.AppendLine("PRINT('------------ Summary ------------')");
-            sqlScript.AppendLine("GO");
-            sqlScript.AppendLine("UPDATE st");
-            sqlScript.AppendLine("SET st.TranslatedText = fd.TranslatedText");
-            sqlScript.AppendLine("FROM X_StaticTranslations_FactoryDefaults fd");
-            sqlScript.AppendLine("LEFT JOIN X_StaticTranslations st on fd.Cd = st.Cd and fd.[Language] = st.LanguageID");
-            sqlScript.AppendLine("WHERE fd.TranslatedText <> st.TranslatedText and isnull(st.NoUpdate,0) = 0;");
-            sqlScript.AppendLine("GO");
-            sqlScript.AppendLine("\tINSERT INTO X_StaticTranslations (Cd, LanguageID, Category, TranslatedText)");
-            sqlScript.AppendLine("SELECT fd.Cd, fd.Language, fd.Category,fd.TranslatedText");
-            sqlScript.AppendLine("FROM X_StaticTranslations_FactoryDefaults fd");
-            sqlScript.AppendLine("LEFT JOIN X_StaticTranslations st on fd.Cd = st.Cd and fd.[Language] = st.LanguageID");
-            sqlScript.AppendLine("WHERE st.Cd is null;");
-            sqlScript.AppendLine("GO");
-            sqlScript.AppendLine("UPDATE o");
-            sqlScript.AppendLine("SET o.VALUE = fd.VALUE");
-            sqlScript.AppendLine("FROM L_Object_FactoryDefaults fd");
-            sqlScript.AppendLine("LEFT JOIN L_Object o on o.ID_TABLE =fd.ID_TABLE and o.ID_LANGUAGES = fd.ID_LANGUAGES and o.TABLE_NAME = fd.TABLE_NAME");
-            sqlScript.AppendLine("WHERE o.VALUE <> fd.VALUE and isnull(o.NO_UPDATE,0) = 0;");
-            sqlScript.AppendLine("GO");
-            sqlScript.AppendLine("\tINSERT INTO L_Object (ID_TABLE, ID_LANGUAGES, DATA_TYPE,VALUE,TABLE_NAME)");
-            sqlScript.AppendLine("SELECT fd.ID_TABLE, fd.ID_LANGUAGES, fd.DATA_TYPE, fd.VALUE, fd.TABLE_NAME");
-            sqlScript.AppendLine("FROM L_Object_FactoryDefaults fd");
-            sqlScript.AppendLine("LEFT JOIN L_Object l on l.ID_TABLE = fd.ID_TABLE and l.ID_LANGUAGES = fd.ID_LANGUAGES and l.TABLE_NAME = fd.TABLE_NAME");
-            sqlScript.AppendLine("WHERE l.ID_TABLE is null AND l.VALUE is null;");
-            sqlScript.AppendLine("GO");
-            sqlScript.AppendLine("PRINT 'Custom Script Completed'");
 
-            
 
             using (StreamWriter files = new StreamWriter(Path.Combine(SQLPath, string.Format("SQLAll.sql")), false, new UTF8Encoding(false)))
             {
-                files.Write(sqlScript);
+                files.Write(sqlScriptFinal);
                 files.Close();
             }
 
@@ -129,29 +144,87 @@ namespace InsertTranslations
             //...
             //3. DataSet - The result of each spreadsheet will be created in the result.Tables
             DataSet result = excelReader.AsDataSet();
+            //6. Free resources (IExcelDataReader is IDisposable)
+            excelReader.Close();
+            //! Move File to Generated Folder
+            stream.Close();
+
             DataTable dt = result.Tables[0];
-            StringBuilder sb = new StringBuilder();
+
             //! Remove First Row Cause its Header
             dt.Rows[0].Delete();
             dt.AcceptChanges();
 
-
-            if (!CdExistsFunction(dt, filePath))
+            if (Path.GetFileName(filePath).Contains("Update"))
             {
-                //! Saved path
-                string SQLPath = Path.Combine(Path.GetFullPath(AppDomain.CurrentDomain.BaseDirectory), "SQL");
+                SaveFile(filePath, dt, true);
+            }
+            else if (!CdExistsFunction(dt, filePath))
+            {
+                SaveFile(filePath, dt, false);
+            }
+        }
 
-                using (StreamWriter file = new StreamWriter(Path.Combine(SQLPath, string.Format(Path.GetFileName(filePath).Split('.')[0] + "_" + DateTime.Now.ToString("yyyyMMdd_HHmm") + ".sql")), false, new UTF8Encoding(false)))
+
+        private static void SaveFile(string filePath, DataTable dt, bool isUpdate)
+        {
+            StringBuilder sb = new StringBuilder();
+            bool hasEmptyValues = false;
+            foreach (DataRow row in dt.Rows)
+            {
+                DataColumnCollection columns = dt.Columns;
+
+                if (isUpdate)
                 {
-                    foreach (DataRow row in dt.Rows)
+                    if (!string.IsNullOrWhiteSpace(row[2].ToString().Trim()) && !string.IsNullOrWhiteSpace(row[1].ToString().Trim()) && !hasEmptyValues)
                     {
-                        DataColumnCollection columns = dt.Columns;
+                        //! Get Mode & Language
+                        int updateMode = Convert.ToInt32(row[2].ToString().Trim());
+                        int languageID = Convert.ToInt32(row[1].ToString().Trim());
 
+                        // Check if setValue is Null or Empty 
+                        if (!string.IsNullOrWhiteSpace(row[3].ToString().Trim())
+                            && (!string.IsNullOrWhiteSpace(row[0].ToString().Trim()) // Check if Cd is Empty for Case 1 or 2
+                            || (!string.IsNullOrWhiteSpace(row[4].ToString().Trim()) && !string.IsNullOrWhiteSpace(row[5].ToString().Trim())))) // check if TableName or ID_Table is empty for case 3
+                        {
+                            //! Update Modes Switch Case
+                            switch (updateMode)
+                            {
+                                // Update Static Translation
+                                case 1:
+                                    sb.AppendFormat("UPDATE X_StaticTranslations_FactoryDefaults SET TranslatedText = N'{0}' WHERE CD = '{1}' AND [Language] = {2}", row[3].ToString().Trim(), row[0].ToString().Trim(), languageID);
+                                    break;
+                                // Update Static Cd
+                                case 2:
+                                    sb.AppendFormat("UPDATE X_StaticTranslations_FactoryDefaults SET [CD] = N'{0}' WHERE CD = '{1}' AND [Language] = {2}", row[3].ToString().Trim(), row[0].ToString().Trim(), languageID);
+                                    break;
+                                // Update Dynamic Translation
+                                case 3:
+                                    sb.AppendFormat("UPDATE L_Object_FactoryDefaults  SET [VALUE] = N'{0}' WHERE TABLE_NAME = '{1}' AND ID_TABLE = {2} AND ID_LANGUAGES = {3}", row[3].ToString().Trim(), row[4].ToString().Trim(), row[5].ToString().Trim(), languageID);
+                                    break;
+                            }
+                            sb.AppendLine("");
+                            sb.AppendLine("GO");
+                        }
+                        else
+                        {
+                            hasEmptyValues = true;
+                        }
+                    }
+                    else
+                    {
+                        hasEmptyValues = true;
+                    }
+                }
+                else
+                {
+                    if (!string.IsNullOrWhiteSpace(row[3].ToString().Trim()) && !hasEmptyValues)
+                    {
                         for (int i = 1; i < 3; i++)
                         {
                             if (row[3].ToString().Length > 0)
-                            {   
-                                sb.AppendFormat("IF NOT EXISTS (select 1 from X_StaticTranslations_FactoryDefaults where [Cd] = '{0}' AND [Language] = {1}) \n", row[3].ToString().Trim().Replace("'","''"), i);
+                            {
+                                sb.AppendFormat("IF NOT EXISTS (select 1 from X_StaticTranslations_FactoryDefaults where [Cd] = '{0}' AND [Language] = {1}) \n", row[3].ToString().Trim().Replace("'", "''"), i);
                                 sb.AppendLine("BEGIN");
                                 sb.AppendFormat("\tINSERT INTO X_StaticTranslations_FactoryDefaults ([Cd], [Language], [TranslatedText], [Category]) VALUES ('{0}', {1}, N'{2}',2) \n", row[3].ToString().Trim().Replace("'", "''"), i, row[i + 3].ToString().Trim().Replace("'", "''"));
                                 sb.AppendLine("END");
@@ -159,21 +232,31 @@ namespace InsertTranslations
                             }
                         }
                     }
-                    sb.AppendLine("");
+                    else
+                    {
+                        hasEmptyValues = true;
+                    }
+                }
+
+            }
+            sb.AppendLine("");
+
+            if (!hasEmptyValues)
+            {
+                //! Saved path and Save to File
+                string SQLPath = Path.Combine(Path.GetFullPath(AppDomain.CurrentDomain.BaseDirectory), "SQL");
+                using (StreamWriter file = new StreamWriter(Path.Combine(SQLPath, string.Format(Path.GetFileName(filePath).Split('.')[0] + "_" + DateTime.Now.ToString("yyyyMMdd_HHmm") + ".sql")), false, new UTF8Encoding(false)))
+                {
                     file.WriteLine(sb.ToString());
                     file.Close();
                 }
-
-                //6. Free resources (IExcelDataReader is IDisposable)
-                excelReader.Close();
-                //! Move File to Generated Folder
-                stream.Close();
                 string GeneratedFolder = Path.Combine(Path.GetFullPath(AppDomain.CurrentDomain.BaseDirectory), "Generated");
                 string destFile = System.IO.Path.Combine(GeneratedFolder, Path.GetFileName(filePath));
                 // Use Path class to manipulate file and directory paths.
                 File.Copy(filePath, destFile, true);
                 File.Delete(filePath);
             }
+            
         }
 
         /// <summary>
